Переносит фильтры в отдельный пакет
