+++ src/main/java/school/faang/user_service/filter/userFilter/UserAboutFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.service;
+package school.faang.user_service.filter.userFilter;
 
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.UserFilterDto;
@@ -7,14 +7,15 @@ import school.faang.user_service.entity.User;
 import java.util.stream.Stream;
 
 @Component
-public class UserPageFilter implements UserFilter {
+public class UserAboutFilter implements UserFilter{
     @Override
     public boolean isApplicable(UserFilterDto filters) {
-        return filters.getPage() > 0 && filters.getPageSize() > 0;
+        return filters.getAboutPattern() != null;
     }
 
     @Override
     public Stream<User> apply(Stream<User> users, UserFilterDto filters) {
-        return users.limit((long) filters.getPageSize() * filters.getPage());
+        return users.filter(user -> user.getAboutMe()
+                .toUpperCase().matches("(.*)%s(.*)".formatted(filters.getAboutPattern().toUpperCase())));
     }
 }
