+++ src/main/java/school/faang/user_service/filter/userFilter/UserCountryFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.service;
+package school.faang.user_service.filter.userFilter;
 
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.UserFilterDto;
@@ -7,15 +7,15 @@ import school.faang.user_service.entity.User;
 import java.util.stream.Stream;
 
 @Component
-public class UserExperienceMinFilter implements UserFilter {
+public class UserCountryFilter implements UserFilter {
     @Override
     public boolean isApplicable(UserFilterDto filters) {
-        return filters.getExperienceMin() > 0;
+        return filters.getCountryPattern() != null;
     }
 
     @Override
     public Stream<User> apply(Stream<User> users, UserFilterDto filters) {
-        return users.filter(user -> user.getExperience() != null
-                && user.getExperience() >= filters.getExperienceMin());
+        return users.filter(user -> user.getCountry().getTitle()
+                .toUpperCase().matches("(.*)%s(.*)".formatted(filters.getCountryPattern().toUpperCase())));
     }
 }
