+++ src/main/java/school/faang/user_service/filter/userFilter/UserPageFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.service;
+package school.faang.user_service.filter.userFilter;
 
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.UserFilterDto;
@@ -7,15 +7,14 @@ import school.faang.user_service.entity.User;
 import java.util.stream.Stream;
 
 @Component
-public class UserPhoneFilter implements UserFilter {
+public class UserPageFilter implements UserFilter {
     @Override
     public boolean isApplicable(UserFilterDto filters) {
-        return filters.getPhonePattern() != null;
+        return filters.getPage() > 0 && filters.getPageSize() > 0;
     }
 
     @Override
     public Stream<User> apply(Stream<User> users, UserFilterDto filters) {
-        return users.filter(user -> user.getPhone()
-                .toUpperCase().matches("(.*)%s(.*)".formatted(filters.getPhonePattern().toUpperCase())));
+        return users.limit((long) filters.getPageSize() * filters.getPage());
     }
 }
