+++ src/main/java/school/faang/user_service/service/SubscriptionService.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Service;
 import school.faang.user_service.dto.UserFilterDto;
 import school.faang.user_service.entity.User;
 import school.faang.user_service.exception.DataValidationException;
+import school.faang.user_service.filter.userFilter.UserFilter;
 import school.faang.user_service.repository.SubscriptionRepository;
 
 import java.util.List;
